refactor: rename scope-member-service to org-member-service
Update all 16 import paths and fix stale test descriptions that
still referenced scopeId or scope_members after the scope-to-org
migration.
