+++ turbo/apps/web/app/api/agent/schedules/[name]/disable/route.ts
@@ -5,7 +5,7 @@ import { getAuthContext } from "../../../../../../src/lib/auth/get-user-id";
 import { disableSchedule } from "../../../../../../src/lib/schedule";
 import { logger } from "../../../../../../src/lib/logger";
 import { isNotFound } from "../../../../../../src/lib/errors";
-import { resolveOrgId } from "../../../../../../src/lib/scope/scope-member-service";
+import { resolveOrgId } from "../../../../../../src/lib/scope/org-member-service";
 
 const log = logger("api:schedules:disable");
