+++ turbo/apps/web/app/api/agent/schedules/[name]/enable/route.ts
@@ -5,7 +5,7 @@ import { getAuthContext } from "../../../../../../src/lib/auth/get-user-id";
 import { enableSchedule } from "../../../../../../src/lib/schedule";
 import { logger } from "../../../../../../src/lib/logger";
 import { isNotFound, isSchedulePast } from "../../../../../../src/lib/errors";
-import { resolveOrgId } from "../../../../../../src/lib/scope/scope-member-service";
+import { resolveOrgId } from "../../../../../../src/lib/scope/org-member-service";
 
 const log = logger("api:schedules:enable");
