+++ turbo/apps/web/app/api/agent/schedules/[name]/route.ts
@@ -12,7 +12,7 @@ import {
 } from "../../../../../src/lib/schedule";
 import { logger } from "../../../../../src/lib/logger";
 import { isNotFound } from "../../../../../src/lib/errors";
-import { resolveOrgId } from "../../../../../src/lib/scope/scope-member-service";
+import { resolveOrgId } from "../../../../../src/lib/scope/org-member-service";
 
 const log = logger("api:schedules:name");
