+++ turbo/apps/web/app/api/cli/auth/test-token/route.ts
@@ -3,7 +3,7 @@ import crypto from "crypto";
 import { initServices } from "../../../../../src/lib/init-services";
 import { cliTokens } from "../../../../../src/db/schema/cli-tokens";
 import { generateDefaultScopeSlug } from "../../../../../src/lib/scope/scope-service";
-import { getDefaultScope } from "../../../../../src/lib/scope/scope-member-service";
+import { getDefaultScope } from "../../../../../src/lib/scope/org-member-service";
 import { orgCache } from "../../../../../src/db/schema/org-cache";
 import { isNotFound } from "../../../../../src/lib/errors";
 import {
