+++ turbo/apps/web/app/api/scope/invite/route.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 import { initServices } from "../../../../src/lib/init-services";
 import { getAuthContext } from "../../../../src/lib/auth/get-user-id";
 import { requireScopeFromRequest } from "../../../../src/lib/scope/resolve-scope";
-import { inviteMember } from "../../../../src/lib/scope/scope-member-service";
+import { inviteMember } from "../../../../src/lib/scope/org-member-service";
 import {
   isBadRequest,
   isNotFound,
