+++ turbo/apps/web/src/lib/run/run-queue-service.ts
@@ -9,7 +9,7 @@ import {
   decryptSecretsMap,
 } from "../crypto/secrets-encryption";
 import { PENDING_RUN_TTL_MS } from "./run-service";
-import { getDefaultScope } from "../scope/scope-member-service";
+import { getDefaultScope } from "../scope/org-member-service";
 import type { CreateRunParams, CreateRunResult } from "./run-service";
 
 const log = logger("service:run-queue");
